ExLink v0.1.0 ExLink.Message

Module used to build Outgoing Messages to send to a Lavalink node.

Types

id() (since 0.1.0)

id() :: String.t() | non_neg_integer()

Represents any discord snowflake.

For example 218348062828003328 or "218348062828003328".

message() (since 0.1.0)

message() :: map()

Represents a built message.

Functions

destroy(guild_id) (since 0.1.0)

destroy(guild_id :: id()) :: message()

Builds a "destroy" message used to tell the server to disconnect from the voice server.

equalizer(bands, guild_id) (since 0.1.0)

equalizer(bands :: list(), guild_id :: id()) :: message()

Builds an "equalizer" message used to change bands of a player.

See this for more info on bands. (You need to scroll down until Using the player equalizer)

pause(paused, guild_id) (since 0.1.0)

pause(paused :: boolean(), guild_id :: id()) :: message()

Builds a "pause" (or "resume") message to pause or resume playback of a player.

play(track, guild_id, opts \\ %{}) (since 0.1.0)

play(
  track :: String.t(),
  guild_id :: id(),
  opts ::
    [startTime: Integer.t(), endTime: Integer.t(), noReplace: boolean()]
    | %{
        optional(:startTime) => Integer.t(),
        optional(:endTime) => Integer.t(),
        optional(:noReplace) => boolean()
      }
) :: message()

Builds a "play" message used to start playing a track.

seek(time_millis, guild_id) (since 0.1.0)

seek(time_millis :: non_neg_integer(), guild_id :: id()) :: message()

Builds a "seek" message used to seek the position of a track.

Time is in milliseconds.

stop(guild_id) (since 0.1.0)

stop(guild_id :: id()) :: message()

Builds a "stop" message used to stop a player.

voice_update(event, session_id, guild_id) (since 0.1.0)

voice_update(event :: map(), session_id :: id(), guild_id :: id()) ::
  message()

Builds a "voice_update" messages used to connect to a voice channel.

event should be a voice_server_update. session_id is obtained from a voice_state_update.

volume(volume, guild_id) (since 0.1.0)

volume(volume :: non_neg_integer(), guild_id :: id()) :: message()

Builds a "volume" message used to change the volume of a player.

Volume may range from 0 to 1000. Default is 100.