ExLink v0.1.0 ExLink.Player behaviour

Player behaviour used to keep track of running players and manage them.

Provides an alternative way using a __using__ macro to start ExLink using the current module as player:

# This time `use ExLink.Player` instead
defmodule MyApp.Player do
  use ExLink.Player

  def init(client, guild_id) do
    {:ok, {client, guild_id}}
  end

  # dispatches not related to a player (for example stats)
  # return value doesn't matter
  def handle_dispatch(data, nil) do
    IO.puts("Players: #{data["players"]}")
    nil
  end

  def handle_dispatch(data, state) do
    IO.puts("Received a #{data["type"]} event")
    {:noreply, state}
  end
end

# Now fits into a supervision tree directly
defmodule MyApp.Supervisor do
  use Supervisor

  def start_link(args) do
    Supervisor.start_link(__MODULE__, args, name: __MODULE__)
  end

  def init(_args) do
    children = [
      {MyApp.Player,
      {%{
          url: "localhost:8080",
          authorization: "123",
          shard_count: 1,
          user_id: 123_456_789_123_456_789,
          # player: Not needed here
        }, name: MyApp.Player}}
    ]

    Supervisor.init(children, strategy: :one_for_one))
  end
end

# after starting the supervisor, play something
message = ExLink.Message.play(track, guild_id)
ExLink.Connection.send(MyApp.Player, message)

Types

common_return() (since 0.1.0)

common_return() ::
  {:noreply, new_state :: term()}
  | {:stop, reason :: term(), new_state :: term()}

Possible return values in handle_info/2, handle_cast/2, and handle_dispatch/2. handle_call/3 also allows to synchronously respond via additional allowed return values.

Functions

call(player, request, timeout \\ 5000) (since 0.1.0)

call(GenServer.server(), term(), timeout()) :: term()

Sends a synchrounous request to the server and waits for its reply. See GenServer.cast/3.

cast(player, request) (since 0.1.0)

cast(GenServer.server(), term()) :: :ok

Sends an asynchrounous request to the server. See GenServer.call/3.

Callbacks

handle_call(request, from, state) (optional) (since 0.1.0)

handle_call(request :: term(), from :: term(), state :: term()) ::
  common_return()
  | {:reply, reply :: term(), new_state :: term()}
  | {:stop, reply :: term(), reason :: term(), new_state :: term()}

Invoked to handle synchronous call/3 messages. call/3 will block until a reply is received (unless the call times out or nodes ar disconnected). See GenServer.handle_call/3.

handle_cast(request, state) (optional) (since 0.1.0)

handle_cast(request :: term(), state :: term()) :: term() :: common_return()

Invoked to handle asynchronous cast/2 messages. See GenServer.handle_cast/2.

handle_dispatch(data, state) (since 0.1.0)

handle_dispatch(data :: map(), state :: term()) :: common_return()

Invoked to handle events from a lavalink node.

For non guild specific dispatches this function will be invoked outside of a player process, with the state nil. Return values will be ignored.

handle_info(msg, state) (optional) (since 0.1.0)

handle_info(msg :: :timeout | term(), state :: term()) :: common_return()

Invoked to handle all other messages. See GenServer.handle_info/2.

init(client, guild_id) (since 0.1.0)

init(client :: pid(), guild_id :: ExLink.Message.id()) ::
  {:ok, term()} | {:stop, reason :: term()}

Invoked when the player is started. See GenServer.init/1.